[#1909] Convert our old dict-based ACLs to list-based ACLs.
This commit includes:

 - addition of ACL and ACE types for model validation
 - replacement of neighborhood ACLs with delegation to the
   --init-- project for that neighborhood
 - unification of security, tool, and delete permission to 'admin'
 - updates to the models
 - updates to permission-checking code
 - rename has_*_access => has_access(*, )
 - removal of some dead code

+++ Allura/allura/controllers/attachments.py
@@ -4,7 +4,7 @@ from webob import exc
 from tg import expose, request, response, redirect
 from ming.utils import LazyProperty
 
-from allura.lib.security import require, has_artifact_access
+from allura.lib.security import require, has_access, require_access
 from .base import BaseController
 
 class AttachmentsController(BaseController):
@@ -28,7 +28,7 @@ class AttachmentController(BaseController):
     edit_perm = 'edit'
 
     def _check_security(self):
-        require(has_artifact_access('read', self.artifact))
+        require_access(self.artifact, 'read')
 
     def __init__(self, filename, artifact):
         self.filename = filename
@@ -55,7 +55,7 @@ class AttachmentController(BaseController):
     @expose()
     def index(self, delete=False, embed=True, **kw):
         if request.method == 'POST':
-            require(has_artifact_access(self.edit_perm, self.artifact))
+            require_access(self.artifact, self.edit_perm)
             if delete:
                 self.attachment.delete()
                 try:
